Update 2.0 supremacyww3.html
Any player who would like to receive free gold in Supremacy WW3 should download the file named "supremacyww3.html" and enter their details as if logging in! This is the biggest update since this website was launched! From now on, you can contact our support without any problems and we are always up to date!